Reference

Reference

Polynomial

StaticPolynomials.Polynomial — Type.

Polynomial(f::MP.AbstractPolynomial, [variables])

Construct a Polynomial from f.

source

StaticPolynomials.coefficients — Function.

coefficients(f)

Return the coefficients of f.

source

StaticPolynomials.exponents — Function.

exponents(::SExponents)

Converts exponents stored in a SExponents to a matrix.

source

exponents(f)

Return the exponents of f as an matrix where each column represents the exponents of a monomial.

source

StaticPolynomials.nvariables — Method.

nvariables(f::Polynomial)

Return the number of variables f.

source

nvariables(F::AbstractSystem)

The number of variables of the system F.

source

StaticPolynomials.coefficienttype — Method.

coefficienttype(f::Polynomial)

Return the type of the coefficients of f.

source

coefficienttype(F::AbstractSystem)

Return the type of the coefficients of the polynomials of F.

source

StaticPolynomials.evaluate — Method.

evaluate(f::Polynomial, x)

Evaluate the polynomial f at x.

source

Base.LinAlg.gradient — Method.

gradient(f::Polynomial, x)

Evaluate the gradient of the polynomial f at x.

source

StaticPolynomials.gradient! — Method.

gradient!(u, f::Polynomial, x)

Evaluate the gradient of the polynomial f at x and store the result in u.

source

StaticPolynomials.evaluate_and_gradient — Function.

evaluate_and_gradient(f::Polynomial, x)

Evaluate the polynomial f and its gradient at x. Returns a tuple.

source

StaticPolynomials.evaluate_and_gradient! — Function.

evaluate_and_gradient!(u, f::Polynomial, x)

Evaluate the polynomial f and its gradient at x. Stores the gradient in u and returns the f(x).

source

Systems of Polynomials

StaticPolynomials.AbstractSystem — Type.

 AbstractSystem{T, M, N}

Represents a system of M polynomials in N variables with coefficients in T.

source

StaticPolynomials.system — Function.

system(polys::AbstractVector{<:MP.AbstractPolynomial}..., variables=sorted_variables(polys))
system(polys...)

Create a system of polynomials from the given polys. The result is an object which is a subtype of AbstractSystem. This function is by design not typestable.

Example

julia> import DynamicPolynomials: @polyvar
julia> @polyvar x y
julia> F = system(x^2+y^2+3, x-y+2, x^2+2y)
julia> F isa AbstractSystem{Int64, 3, 2}
true

source

StaticPolynomials.nvariables — Function.

nvariables(f::Polynomial)

Return the number of variables f.

source

nvariables(F::AbstractSystem)

The number of variables of the system F.

source

StaticPolynomials.npolynomials — Function.

npolynomials(F::AbstractSystem)

The number of polynomials of the system F.

source

StaticPolynomials.coefficienttype — Function.

coefficienttype(f::Polynomial)

Return the type of the coefficients of f.

source

coefficienttype(F::AbstractSystem)

Return the type of the coefficients of the polynomials of F.

source

StaticPolynomials.evaluate — Method.

evaluate(F::AbstractSystem, x::AbstractVector)

Evaluate the system F at x.

source

StaticPolynomials.evaluate! — Method.

evaluate!(u, F::AbstractSystem, x)

Evaluate the system F at x and store the result in u.

source

StaticPolynomials.jacobian — Function.

jacobian(F::AbstractSystem, x)
jacobian(F::AbstractSystem, x::SVector)

Evaluate the Jacobian of the system F at x.

source

StaticPolynomials.jacobian! — Function.

jacobian!(u, F::AbstractSystem, x)

Evaluate the Jacobian of the system F at x and store the result in u.

source

StaticPolynomials.evaluate_and_jacobian — Function.

evaluate_and_jacobian(F::AbstractSystem, x)
evaluate_and_jacobian(F::AbstractSystem, x::SVector)

Evaluate the system F and its Jacobian at x.

source

StaticPolynomials.evaluate_and_jacobian! — Function.

evaluate_and_jacobian!(u, U, F::AbstractSystem, x)

Evaluate the system F and its Jacobian at x and store the results in u (evalution) and U (Jacobian).

source