Reference

Reference

Polynomial

StaticPolynomials.Polynomial — Type.

Polynomial{T, SE<:SExponents}

A Polynomial with coefficents in T in NVars variables and exponents of type SE.

Polynomial(f::MP.AbstractPolynomial, [variables])

Construct a Polynomial from f.

source

StaticPolynomials.SExponents — Type.

SExponents{N}

Store the exponents of the terms of a polynomial (the support) with N terms as an type. This results in an unique type for each possible support.

source

StaticPolynomials.coefficients — Function.

coefficients(f)

Return the coefficients of f.

source

StaticPolynomials.exponents — Function.

exponents(::SExponents)

Converts exponents stored in a SExponents to a matrix.

source

exponents(f)

Return the exponents of f as an matrix where each column represents the exponents of a monomial.

source

StaticPolynomials.nvariables — Method.

nvariables(f::Polynomial)

Return the number of variables of f.

source

nvariables(F::AbstractSystem)

The number of variables of the system F.

source

StaticPolynomials.nparameters — Method.

nparameters(f::Polynomial)

Return the number of parameters of f.

source

StaticPolynomials.coefficienttype — Method.

coefficienttype(f::Polynomial)

Return the type of the coefficients of f.

source

StaticPolynomials.scale_coefficients! — Method.

scale_coefficients!(f, λ)

Scale the coefficients of f by the factor λ.

source

StaticPolynomials.evaluate — Method.

evaluate(f::Polynomial, x)

Evaluate the polynomial f at x.

source

StaticPolynomials.evaluate — Method.

evaluate(f::Polynomial, x, p)

Evaluate f at x with parameters p.

source

LinearAlgebra.gradient — Function.

gradient(f::Polynomial, x)

Evaluate the gradient of the polynomial f at x.

source

gradient(f::Polynomial, x, p)

Evaluate the gradient of the polynomial f at x with parameters p.

source

StaticPolynomials.gradient! — Function.

gradient!(u, f::Polynomial, x)

Evaluate the gradient of the polynomial f at x and store the result in u.

gradient!(u, f::Polynomial, x, p)

Evaluate the gradient of the polynomial f at x with parameters p and store the result in u.

source

StaticPolynomials.evaluate_and_gradient — Function.

evaluate_and_gradient(f::Polynomial, x)

Evaluate the polynomial f and its gradient at x. Returns a tuple.

evaluate_and_gradient(f::Polynomial, x, p)

Evaluate the polynomial f and its gradient at x with parameters p. Returns a tuple.

source

StaticPolynomials.evaluate_and_gradient! — Function.

evaluate_and_gradient!(u, f::Polynomial, x)

Evaluate the polynomial f and its gradient at x. Stores the gradient in u and returns the f(x).

evaluate_and_gradient!(u, f::Polynomial, x, p)

Evaluate the polynomial f and its gradient at x with parameters p. Stores the gradient in u and returns the f(x).

source

StaticPolynomials.differentiate_parameters — Method.

differentiate_parameters(f::Polynomial, x, p)

Evaluate the gradient of the polynomial f w.r.t. the parameters at x with parameters p.

source

StaticPolynomials.differentiate_parameters! — Method.

differentiate_parameters!(u, f::Polynomial, x, p)

Evaluate the gradient of the polynomial f w.r.t. the parameters at x with parameters p and store the result in u.

source

Systems of Polynomials

StaticPolynomials.PolynomialSystem — Type.

PolynomialSystem{N, NVars, NParams, <:Tuple}

A polynomial system of N polynomials in NVars variables with NParams variables.

Constructors:

PolynomialSystem(polys::AbstractVector{<:MP.AbstractPolynomial}; variables=MP.variables(polys), parameters=nothing)
PolynomialSystem(polys::MP.AbstractPolynomial...; kwargs...)

Create a system of polynomials from the given polynomials polys. This function is by design not typestable.

Example

julia> import DynamicPolynomials: @polyvar
julia> @polyvar x y a
julia> PolynomialSystem(x^2+y^2+3, x-y+2, x^2+2y+a; parameters=[a])
PolynomialSystem{3, 2}:
 3 + x² + y²

 2 + x - y

 x² + 2y

source

StaticPolynomials.npolynomials — Method.

npolynomials(F::AbstractSystem)

The number of polynomials of the system F.

source

StaticPolynomials.nvariables — Method.

nvariables(f::Polynomial)

Return the number of variables of f.

source

nvariables(F::AbstractSystem)

The number of variables of the system F.

source

StaticPolynomials.nparameters — Method.

nparameters(F::AbstractSystem)

The number of parameters of the system F.

source

Base.foreach — Method.

foreach(f, F::AbstractSystem)

Iterate over the polynomials of F and apply f to each polynomial.

source

StaticPolynomials.scale_coefficients! — Method.

scale_coefficients!(F::AbstractSystem{T, M}, λ::AbstractVector)

Scale the coefficients of the polynomials fᵢ of F by the factor λᵢ. λ needs to have have length M.

source

StaticPolynomials.evaluate — Method.

 evaluate(F::PolynomialSystem, x)

Evaluate the polynomial system F at x.

source

StaticPolynomials.evaluate — Method.

 evaluate(F::PolynomialSystem, x, p)

Evaluate the polynomial system F at x with parameters p.

source

StaticPolynomials.evaluate! — Method.

 evaluate!(u, F::PolynomialSystem, x)

Evaluate the polynomial system F at x and store its result in u.

source

StaticPolynomials.evaluate! — Method.

 evaluate!(u, F::PolynomialSystem, x, p)

Evaluate the polynomial system F at x with parameters p and store its result in u.

source

StaticPolynomials.jacobian — Function.

 jacobian(F::PolynomialSystem, x)

Evaluate the Jacobian of the polynomial system F at x.

source

 jacobian(F::PolynomialSystem, x, p)

Evaluate the Jacobian of the polynomial system F at x with parameters p.

source

StaticPolynomials.jacobian! — Function.

 jacobian!(U, F::PolynomialSystem, x)

Evaluate the Jacobian of the polynomial system F at x and store its result in U.

source

 jacobian(U, F::PolynomialSystem, x, p)

Evaluate the Jacobian of the polynomial system F at x with parameters p and store its result in U.

source

StaticPolynomials.evaluate_and_jacobian — Function.

evaluate_and_jacobian!(F::PolynomialSystem, x, p)

Evaluate the system F and its Jacobian at x with parameters p.

source

StaticPolynomials.evaluate_and_jacobian! — Function.

evaluate_and_jacobian!(u, U, F::PolynomialSystem, x)

Evaluate the system F and its Jacobian at x and store the results in u (evalution) and U (Jacobian).

source

evaluate_and_jacobian!(u, U, F::PolynomialSystem, x, p)

Evaluate the system F and its Jacobian at x with parameters p and store the results in u (evalution) and U (Jacobian).

source

evaluate_and_jacobian(F::PolynomialSystem, x)

Evaluate the system F and its Jacobian at x.

source

StaticPolynomials.differentiate_parameters — Method.

 differentiate_parameters(F::PolynomialSystem, x, p)

Evaluate the derivative of the polynomial system F at x with parameters p with respect to the parameters and store the result in U.

source

StaticPolynomials.differentiate_parameters! — Method.

 differentiate_parameters!(U, F::PolynomialSystem, x, p)

Evaluate the derivative of the polynomial system F at x with parameters p with respect to the parameters and store the result in U.

source